qa/tasks/workunit: factor out overrides and refspec logic
Allow for reuse in the cram task.

(cherry picked from commit e1c89b51c80407fadbfee82d5d396cfcbd791aae)
